Fixes for failing test cases
GitHub proofs were failing because the CSS classes had changed. In addition, BeautifulSoup's html.body.find() was not working as expected (changed to html.find())

A Facebook proof was failing because the test was looking for the Username in the URL. The username had a dot that was being redirect by Facebook to the dotless version. Updated the check to allow dotted or dotless username in the Facebook url.